Introduction and use of text editors
This is a program module that allows you to embed an editor for your information on your site:

To install this module, you can enter the following command in the terminal:

pip install django-summernote

After installation, you should enter "django_summernote" in your installed programs in Django site settings and django_summernote.urls in urls.py and make sure that MEDlA_URL is set well and finally a migrate should be done.

For more information, you can refer to the following addresses.

https://github.com/summernote/django-summernote

Keep in mind that there are other modules for this, such as Ckeditor and tinmyce and many other things that can be used in pages.

     https://github.com/django-ckeditor/django-ckeditor
     https://github.com/jazzband/django-tinymce